launch: handle type "yaml" parameters (new in roslaunch since Lunar)
This is actually a bit complicated, since this breaks a previous assumption
we made: Our lazy evaluation of parameters depend on a 1:1 mapping of
parameter names to jobs - this is not the case with YAML parameters, since
one YAML file can turn into multiple params on the parameter server.

So we handle YAML parameters separately from "ordinary" parameters, i.e.
here our lazy evaluation does not prevent multiple loadings of the same
parameters.