test/xml: avoid multiline Catch captures in exception tests
These trigger some weird bug between g++ 5.4 and ccache, which is used
in the ROS buildfarm. Apparently, when g++ is invoked by ccache, it gets
the option `-save-temps` or similar. In that case, the preprocessor
outputs invalid C++ code for the stringified multiline expressions.

Workaround for now: Don't capture these directly, instead capture a
variable containing the string.

This should fix compilation on the build farm.