gui: Fix index bug in showContextMenu
Fix a bug that could lead to performing start/stop actions on wrong nodes.
This happened because the index used to determine the node those actions
should be performed on was the one returned by the view, which is the
index of the proxy model. This may be different than the one of the real
model.