launch: launch_config: Add stop timeout attribute
Add support for rosmon-stop-timeout node attribute. This attribute can
be used to define the seconds to wait when a stop request is made,
before forcing the node to stop.

If the rosmon-stop-timeout attribute is not present, a default timeout
is used. The default timeout is set to be 5 seconds, but can be changed
using the LauchConfig::setDefaultStopTimeout() method.