rosmon_core: monitor: remove superfluous `exists` check for /proc/.../stat
We use boost::filesystem to check if the stat file exists. It seems that
`fs::exists` handles ESRCH (which procfs may return in some cases, see #81)
as an exception, which is then unhandled by us.

Instead of catching the exception, we can just remove the check since
`readStatFile()` handles any error on opening the stat file correctly and
causes a `continue` in the process iteration loop.